Fixes for new tmk_core code
Fix right modifier handling in code: Use keycode mods instead of the MOD_* constants, which I apparently shouldn't have been using in the first place.

Also overload MEDIA_NEXT so it is MEDIA_PREV when shifted

Also fix anykey let-up problems

TMK's fix for stuck keys (with the NO_TRACK_KEY_PRESS macro) makes it so that keys pressed on one layer are let-up on that layer even if the layer's changed in the meantime. This is the right way of doing things, but it invalidates some of my assumptions, namely that the active layer will always be the one that a key press originated in.

The let-up problem is that when doing out-of-order releases with an anykey, the anykey function gets called for the key, but the layer doesn't match that key anymore, so the function doesn't know what to do. (Previously, this was caught by the catch-all "clear everything but mods" when changing layers.)

Fix it by putting a parameter in the fn_action for anykey with the correct home layer for the key. This makes anykey take one fn_action per layer, but it's the right way of doing things, and I have actions to spare.